news-curator/requirements.txt
# News Curator Agent Dependencies
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
faker==19.6.2
openai==1.3.0